Institution Management System with updated UI
