+++ Python Core Foundations/Exceptions, Files & JSON/Data/announcements.json
@@ -1,10 +1,1 @@
-[
-  {
-    "id": 1,
-    "title": "hi",
-    "content": "sdfvesd",
-    "posted_at": "2026-02-05 17:09:33",
-    "deadline": "2026-02-06",
-    "visible": true
-  }
-]
+[]
